Rewrite, simplify, and harden parser
This rewrites the vanilla parser to use ByteBuffer slices. This
makes the parsing code a lot easier to reason about and also makes it
robust to cases that would previously have caused the implementation
to raise internal assertion failures.

In addition, a test factory has been added that subjects the parser
to around ~10000 randomly corrupted RIFF files in order to verify
that internal assertions are never triggered.